ai_working/amplifier-v2/amplifier-dev/amplifier/pyproject.toml
[project]
name = "amplifier"
version = "0.1.0"
description = "User-facing CLI tool for the Amplifier system"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Amplifier Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "amplifier-core @ file:///home/brkrabac/repos/amplifier.v2-exploration/ai_working/amplifier-v2/amplifier-dev/amplifier-core",
    "click>=8.1.7",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
]

[project.scripts]
amplifier = "amplifier.cli:main"
amp = "amplifier.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=6.1.1",
    "pytest-mock>=3.14.0",
    "ruff>=0.11.10",
    "pyright>=1.1.405",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["amplifier"]

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = [
    "**/__pycache__",
    ".venv/**",
    ".git/**",
    ".ruff_cache/**",
    ".pytest_cache/**",
    "*.egg-info",
    "build/**",
    "dist/**",
]
typeCheckingMode = "basic"
reportMissingImports = false
pythonVersion = "3.11"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["amplifier"]
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]